backend/pyproject.toml
[project]
name = "finquery"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "camelot-py[cv]>=1.0.9",
    "chromadb>=1.3.5",
    "fastapi>=0.124.2",
    "psycopg2-binary>=2.9.9",
    "sqlalchemy>=2.0.25",
    "langchain>=1.1.3",
    "langchain-text-splitters>=1.0.0",
    "openai>=2.8.1",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pymupdf-layout>=1.26.6",
    "pymupdf4llm>=0.2.7",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.2.0",
    "tabulate>=0.9.0",
    "tiktoken>=0.12.0",
    "together>=1.5.32",
    "uvicorn>=0.38.0",
]
